move migration and knexfile.js to root to run migration automatically
