+++ package.json
@@ -14,8 +14,8 @@
     "lint": "eslint .",
     "lintfix": "eslint . --fix",
     "postinstall": "yarn build",
-    "migrate": "knex migrate:latest --knexfile server/knexfile.js",
-    "rollback": "knex migrate:rollback --knexfile server/knexfile.js"
+    "migrate": "knex migrate:latest --knexfile knexfile.js",
+    "rollback": "knex migrate:rollback --knexfile knexfile.js"
   },
   "dependencies": {
     "body-parser": "~1.17.1",
