+++ server/scripts/populate-event-img.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs');
 const knex = require('knex');
-const knexConfig = require('../knexfile');
+const knexConfig = require('../../knexfile');
 
 const db = knex(knexConfig);
